// packages/admin-antd/src/types/model/dealEntity.ts
export enum DealType {
	ALL = "",
	PARTNERSHIPS = "partnerships",
	CAPITAL_RAISING = "capital_raising",
	STARTUP_PITCH = "startup_pitch",
	EQUITY = "equity",
	SELL_A_BUSINESS = "sell_a_business",
}